move majority of SSD1680 and SSD1681 driver code into the SSD168x base class
