+++ devices/ePaper/Drivers/Ssd168x/Ssd168x.cs
@@ -8,7 +8,6 @@ using System.Drawing;
 using System.Threading;
 
 using Iot.Device.EPaper.Buffers;
-using Iot.Device.EPaper.Drivers.Ssd168x.Ssd1681;
 using Iot.Device.EPaper.Enums;
 using Iot.Device.EPaper.Primitives;
 
@@ -27,6 +26,16 @@ namespace Iot.Device.EPaper.Drivers.Ssd168x
 
         private bool _shouldDispose;
         private bool _disposed;
+
+        /// <summary>
+        /// The max supported clock frequency for the SSD168x controller. 20MHz.
+        /// </summary>
+        public const int SpiClockFrequency = 20_000_000;
+
+        /// <summary>
+        /// The supported <see cref="System.Device.Spi.SpiMode"/> by the SSD168x controller.
+        /// </summary>
+        public const SpiMode SpiMode = System.Device.Spi.SpiMode.Mode0;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="Ssd168x"/> class.
@@ -53,16 +62,6 @@ namespace Iot.Device.EPaper.Drivers.Ssd168x
             bool enableFramePaging = true,
             bool shouldDispose = true)
         {
-            if (width <= 0 || width > 200)
-            {
-                throw new ArgumentOutOfRangeException(nameof(width));
-            }
-
-            if (height <= 0 || height > 200)
-            {
-                throw new ArgumentOutOfRangeException(nameof(height));
-            }
-
             this._spiDevice = spiDevice ?? throw new ArgumentNullException(nameof(spiDevice));
             this._gpioController = gpioController ?? new GpioController();
 
@@ -90,7 +89,18 @@ namespace Iot.Device.EPaper.Drivers.Ssd168x
         public virtual int Height { get; protected set; }
 
         /// <inheritdoc/>
-        public abstract IFrameBuffer FrameBuffer { get; protected set; }
+        public virtual IFrameBuffer FrameBuffer
+        {
+            get
+            {
+                return FrameBuffer2bpp;
+            }
+
+            protected set
+            {
+                FrameBuffer2bpp = (FrameBuffer2BitPerPixel)value;
+            }
+        }
 
         /// <inheritdoc/>
         public virtual bool PagedFrameDrawEnabled { get; protected set; }
@@ -129,6 +139,11 @@ namespace Iot.Device.EPaper.Drivers.Ssd168x
         /// Gets the number of pages in every frame buffer.
         /// </summary>
         protected abstract int PagesPerFrame { get; }
+
+        /// <summary>
+        /// Gets or sets the <see cref="FrameBuffer2BitPerPixel"/> used internally by <see cref="Ssd168x"/> devices to represents the frame.
+        /// </summary>
+        protected FrameBuffer2BitPerPixel FrameBuffer2bpp { get; set; }
 
         /// <summary>
         /// Gets the index of the first frame page.
@@ -192,8 +207,124 @@ namespace Iot.Device.EPaper.Drivers.Ssd168x
             }
         }
 
+        /// <summary>
+        /// Draws the specified buffer directly to the Black/White RAM on the display.
+        /// Call <see cref="PerformFullRefresh"/> after to update the display.
+        /// </summary>
+        /// <param name="startXPos">X start position.</param>
+        /// <param name="startYPos">Y start position.</param>
+        /// <param name="buffer">The buffer array to draw.</param>
+        public void DirectDrawBuffer(int startXPos, int startYPos, params byte[] buffer)
+        {
+            SetPosition(startXPos, startYPos);
+
+            SendCommand((byte)Command.WriteBackWhiteRAM);
+            SendData(buffer);
+        }
+
+        /// <summary>
+        /// Draws the specified buffer directly to the Red RAM on the display.
+        /// Call <see cref="PerformFullRefresh"/> after to update the display.
+        /// </summary>
+        /// <param name="startXPos">X start position.</param>
+        /// <param name="startYPos">Y start position.</param>
+        /// <param name="buffer">The buffer array to draw.</param>
+        public void DirectDrawColorBuffer(int startXPos, int startYPos, params byte[] buffer)
+        {
+            SetPosition(startXPos, startYPos);
+
+            SendCommand((byte)Command.WriteRedRAM);
+            SendData(buffer);
+        }
+
+        /// <summary>
+        /// Draws a single pixel to the appropriate frame buffer.
+        /// </summary>
+        /// <param name="x">The X Position.</param>
+        /// <param name="y">The Y Position.</param>
+        /// <param name="inverted">True to invert the pixel from white to black.</param>
+        /// <remarks>
+        /// The SSD168x comes with 2 RAMs: a Black and White RAM and a Red RAM.
+        /// Writing to the B/W RAM draws B/W pixels on the panel. While writing to the Red RAM draws red pixels on the panel (if the panel supports red).
+        /// However, the SSD168x doesn't support specifying the color level (no grayscaling), therefore the way the buffer is selected 
+        /// is by performing a simple binary check: 
+        /// if R >= 128 and G == 0 and B == 0 then write a red pixel to the Red Buffer/RAM
+        /// if R == 0 and G == 0 and B == 0 then write a black pixel to B/W Buffer/RAM
+        /// else, assume white pixel and write to B/W Buffer/RAM.
+        /// </remarks>
+        public void DrawPixel(int x, int y, bool inverted)
+        {
+            DrawPixel(x, y, inverted ? Color.Black : Color.White);
+        }
+
         /// <inheritdoc/>
-        public abstract void DrawPixel(int x, int y, Color color);
+        /// <remarks>
+        /// The SSD168x comes with 2 RAMs: a Black and White RAM and a Red RAM.
+        /// Writing to the B/W RAM draws B/W pixels on the panel. While writing to the Red RAM draws red pixels on the panel (if the panel supports red).
+        /// However, the SSD168x doesn't support specifying the color level (no grayscaling), therefore the way the buffer is selected 
+        /// is by performing a simple binary check: 
+        /// if R >= 128 and G == 0 and B == 0 then write a red pixel to the Red Buffer/RAM
+        /// if R == 0 and G == 0 and B == 0 then write a black pixel to B/W Buffer/RAM
+        /// else, assume white pixel and write to B/W Buffer/RAM.
+        /// </remarks>
+        public virtual void DrawPixel(int x, int y, Color color)
+        {
+            // ignore out of bounds draw attempts
+            if (x < 0 || x >= Width || y < 0 || y >= Height)
+            {
+                return;
+            }
+
+            var frameByteIndex = GetFrameBufferIndex(x, y);
+            var pageByteIndex = frameByteIndex - CurrentFrameBufferPageLowerBound;
+
+            // if the specified point falls in the current page, update the buffer
+            if (CurrentFrameBufferPageLowerBound <= frameByteIndex
+                && frameByteIndex < CurrentFrameBufferPageUpperBound)
+            {
+                /*
+                 * Lookup Table for colors on SSD168x
+                 * 
+                 *  LUT for Red, Black, and White ePaper display
+                 * 
+                 * |                |               |                    |
+                 * | Data Red RAM   | Data B/W RAM  | Result Pixel Color |
+                 * |----------------|---------------|--------------------|
+                 * |        0       |       0       |       Black        |
+                 * |        0       |       1       |       White        |
+                 * |        1       |       0       |       Red          |
+                 * |        1       |       1       |       Red          |
+                 * 
+                 * 
+                 *  LUT for Black and White ePaper display with SSD168x
+                 * |                |               |                    |
+                 * | Data Red RAM   | Data B/W RAM  | Result Pixel Color |
+                 * |----------------|---------------|--------------------|
+                 * |        0       |       0       |       Black        |
+                 * |        0       |       1       |       White        |
+                 * |        1       |       0       |       Black        |
+                 * |        1       |       1       |       White        |
+                 */
+
+                if (color.R >= 128 && color.G == 0 && color.B == 0)
+                {
+                    // this is a colored pixel
+                    // according to LUT, no need to change the pixel value in B/W buffer.
+                    // red frame buffer starts with 0x00. ORing it to set red pixel to 1.
+                    FrameBuffer2bpp.ColorBuffer[pageByteIndex] |= (byte)(128 >> (x & 7));
+                }
+                else if (color.R == 0 && color.G == 0 && color.B == 0)
+                {
+                    // black pixel
+                    FrameBuffer2bpp.BlackBuffer[pageByteIndex] &= (byte)~(128 >> (x & 7));
+                }
+                else
+                {
+                    // assume white if R, G, and B > 0
+                    FrameBuffer2bpp.BlackBuffer[pageByteIndex] |= (byte)(128 >> (x & 7));
+                }
+            }
+        }
 
         /// <inheritdoc />
         public virtual void EndFrameDraw()
@@ -213,7 +344,19 @@ namespace Iot.Device.EPaper.Drivers.Ssd168x
         }
 
         /// <inheritdoc/>
-        public abstract void Flush();
+        public virtual void Flush()
+        {
+            // write B/W and Color (RED) frame to the display's RAM.
+            DirectDrawBuffer(
+                CurrentFrameBufferStartXPosition,
+                CurrentFrameBufferStartYPosition,
+                FrameBuffer2bpp.BlackBuffer.Buffer);
+
+            DirectDrawColorBuffer(
+                CurrentFrameBufferStartXPosition,
+                CurrentFrameBufferStartYPosition,
+                FrameBuffer2bpp.ColorBuffer.Buffer);
+        }
 
         /// <summary>
         /// Gets the index of the byte containing the pixel specified by the <paramref name="x"/> and <paramref name="y"/> parameters.
@@ -283,7 +426,48 @@ namespace Iot.Device.EPaper.Drivers.Ssd168x
         /// Perform the required initialization steps to set up the display.
         /// </summary>
         /// <exception cref="InvalidOperationException">Unable to initialize the display until it has been powered on. Call PowerOn() first.</exception>
-        public abstract void Initialize();
+        public virtual void Initialize()
+        {
+            if (PowerState != PowerState.PoweredOn)
+            {
+                throw new InvalidOperationException();
+            }
+
+            // set gate lines and scanning sequence
+            SendCommand((byte)Command.DriverOutputControl);
+
+            // refer to the datasheet for a description of the parameters
+            SendData((byte)Height, 0x00, 0x00);
+
+            // Set data entry sequence
+            SendCommand((byte)Command.DataEntryModeSetting);
+
+            // Y Increment, X Increment with RAM address counter auto incremented in the X direction.
+            SendData(0x03);
+
+            // Set RAM X start / end position
+            SendCommand((byte)Command.SetRAMAddressXStartEndPosition);
+
+            // Param1: Start at 0 | Param2: End at display width converted to bytes
+            SendData(0x00, (byte)((Width / 8) - 1));
+
+            // Set RAM Y start / end positon
+            SendCommand((byte)Command.SetRAMAddressYStartEndPosition);
+
+            // Param1 & 2: Start at 0 | Param3 & 4: End at display height converted to bytes
+            SendData(0x00, 0x00, (byte)(Height - 1), 0x00);
+
+            // Set Panel Border
+            SendCommand((byte)Command.BorderWaveformControl);
+            SendData(0xc0);
+
+            // Set Temperature sensor to use internal temp sensor
+            SendCommand((byte)Command.TempSensorControlSelection);
+            SendData(0x80);
+
+            // Do a full refresh of the display
+            PerformFullRefresh();
+        }
 
         /// <summary>
         /// Initializes a new instance of the intenral frame buffer. Supports paging.
@@ -291,7 +475,11 @@ namespace Iot.Device.EPaper.Drivers.Ssd168x
         /// <param name="width">The width of the frame buffer in pixels.</param>
         /// <param name="height">The height of the frame buffer in pixels.</param>
         /// <param name="enableFramePaging">If <see langword="true"/>, enables paging the frame.</param>
-        protected abstract void InitializeFrameBuffer(int width, int height, bool enableFramePaging);
+        protected virtual void InitializeFrameBuffer(int width, int height, bool enableFramePaging)
+        {
+            var frameBufferHeight = enableFramePaging ? height / PagesPerFrame : height;
+            FrameBuffer2bpp = new FrameBuffer2BitPerPixel(frameBufferHeight, width);
+        }
 
         /// <inheritdoc/>
         public virtual bool NextFramePage()
@@ -393,7 +581,18 @@ namespace Iot.Device.EPaper.Drivers.Ssd168x
         /// to persist the frame to the display's RAM before calling this method.
         /// </summary>
         /// <param name="page">The frame buffer page index to move to.</param>
-        protected abstract void SetFrameBufferPage(int page);
+        protected virtual void SetFrameBufferPage(int page)
+        {
+            if (page < 0 || page >= PagesPerFrame)
+            {
+                page = 0;
+            }
+
+            FrameBuffer2bpp.Clear();
+            CurrentFrameBufferPage = page;
+
+            CalculateFrameBufferPageBounds();
+        }
 
         /// <inheritdoc/>
         public virtual void SetPosition(int x, int y)
